requirements.txt
alembic==1.7.5
aniso8601==9.0.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
